migrated import contacts controller to typescript
