+++ backend/src/routes/contactRoutes.ts
@@ -2,11 +2,15 @@ import express from "express";
 import isAuth from "../middleware/isAuth";
 
 import * as ContactController from "../controllers/ContactController";
-// import ImportPhoneContactsController from "../controllers/ImportPhoneContactsController";
+import * as ImportPhoneContactsController from "../controllers/ImportPhoneContactsController";
 
 const contactRoutes = express.Router();
 
-// contactRoutes.post("/contacts/import", isAuth, ImportPhoneContactsController.store);
+contactRoutes.post(
+  "/contacts/import",
+  isAuth,
+  ImportPhoneContactsController.store
+);
 
 contactRoutes.get("/contacts", isAuth, ContactController.index);
